vimrc: remove 'call add(plugins..)' system, just use 'Plug' instead
I had been using this as a way to automatically PlugInstall any plugged
plugins that weren't installed yet. The 'plugins' list is iterated
through once all plugins are declared and 'Plug'-ed, the same list is
then compared to the './vim/plugged' dir contents to check for
uninstalled plugins.

But this list is created for us already, the 'g:plugins_order' variable.
This means we don't have to build our own list, and we can just Plug
the plugins straight away.

This also means we can use 'Plug's extended syntax, e.g. the
post-update hook, which I'm using to 'fzf#install' the 'junegunn/fzf'
plugin as its docs suggest, ensuring the latest fzf binary is installed.