zshrc: alias 'rg' to 'rg --no-ignore-vcs'
I'm not a fan of rg using .gitignore, just because I'm not tracking a
file in git, doesn't mean I wouldn't like to search through it.

But this was particularly bothering me because, after a recent update,
rg was going travelling all the way to my home dir's .gitignore when I
was searching in a non-git directory. And that .gitignore ignores
*everything* other than selected dotfiles.