vimrc: show last-opened buffer in status line + map for opening it
This new map is preferable to the '#' solution I had before, since '#'
can refer to a buffer even after it's been closed, which messes with
my workflow. Instead, this new <S-Tab> map will take me to the last-open
buffer that is still 'listed', i.e. is still open.

Also added a WIP 'LightlineRecentBufs' function that shows a list
of all open buffers, sorted by when they were last open. I experimented
with showing this in the statusline + tabline, but both didn't feel
right. Still, I'm going to keep it around for a bit.