vimrc: Switch to # buffer on S-Tab
Before this S-Tab switched back to the most-recently-open buffer
according to FZF (which it tracks for us). But the advantage of #
over this is that it'll track each window's last-opened buffer
separately. FZF's most-recently-open buffer is the same for all
windows.

# can refer to deleted buffers though which isn't ideal. I've left
a TODO for a potential workaround for this.

Also removed that 'LightlineRecentBufs' function. I'm not going
to use it.