vimrc: have <Space>* set FZF query, vs 'prefiltering'
Passing arguments straight to :Rg tells FZF to run rg with the given
pattern, then let us search through those the results interactively.

Now, <Space>* just launches an empty :Rg search, and adds the word under
the cursor to the window, as though we'd typed it in. This is preferable
- it lets us edit the search term, vs being stuck searching a
pre-filtered results set.