vimrc: add ctrl-y 'expect' to Rg commands
Hitting ctrl-y when using any of the Rg commands will now yank the
chosen line into the default yank register (").

I had to set the sink function as a 'sink*' option, which is intended to
be used when there are multiple ways of 'completing' an FZF command. The
output sink* functions receive is a tuple array, containing the key used
to complete FZF, and the regular FZF output (in this case, the filename
and line number), e.g.

`[ 'ctrl-y', '.vimrc:7:"foo' ]`

To register ctrl-y as a way to complete FZF, we use the '--expect' flag,
given to the underlying FZF call. (When completing with <Enter>, that
first element in the above array is ''.)